//修改主键策略，权限

+++ src/main/java/com/fux/afk/auth/repository/RolePermissionRepository.java
@@ -3,11 +3,11 @@ package com.fux.afk.auth.repository;
 import com.fux.afk.auth.entity.SysRolePermission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
+
 /**
  * Created by fuxj on 2019-4-4
  */
-public interface RolePermissionRepository extends JpaRepository<SysRolePermission, Integer> {
-
-
+public interface RolePermissionRepository extends JpaRepository<SysRolePermission, BigDecimal> {
 
 }
