+++ src/main/java/com/fux/afk/auth/repository/UserRepository.java
@@ -4,10 +4,12 @@ import com.fux.afk.auth.entity.SysUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
+
 /**
  * Created by fuxj on 2019-4-2
  */
-public interface UserRepository extends JpaRepository<SysUser, Integer> {
+public interface UserRepository extends JpaRepository<SysUser, BigDecimal> {
 
     @Query(value = "select * from sys_user where name = ?", nativeQuery = true)
     SysUser getUserByName(String name);
