+++ src/main/java/com/fux/afk/auth/service/PermissionService.java
@@ -3,6 +3,7 @@ package com.fux.afk.auth.service;
 import com.fux.afk.auth.entity.SysPermission;
 import com.fux.afk.support.vo.ResultVo;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -10,13 +11,13 @@ import java.util.List;
  */
 public interface PermissionService {
 
-    List<SysPermission> getListByRoleId(Integer roleId);
+    List<SysPermission> getListByRoleId(BigDecimal roleId);
 
     List<SysPermission> list();
 
-    SysPermission getPermissionById(Integer id);
+    SysPermission getPermissionById(BigDecimal id);
 
     ResultVo saveOrUpdate(SysPermission permission);
 
-    ResultVo delete(Integer id);
+    ResultVo delete(BigDecimal id);
 }
